Public API

PlutoVista.VTKPlot — Type

Structure containig plot information. In particular it contains dict of data sent to javascript.

PlutoVista.VTKPlot — Method

    VTKPlot(;resolution=(300,300))

Create a canvas plot with given resolution in the notebook and given "world coordinate" range.

PlutoVista.axis2d! — Method

axis2d!(vtkplot)

Add 2D coordinate system axes to the plot. Sets camera handling to 2D mode.

PlutoVista.axis3d! — Method

axis3d!(vtkplot)

Add 3D coordinate system axes to the plot. Sets camera handling to 3D mode.

PlutoVista.tricontour! — Method

 tricontour!(p::VTKPlot,pts, tris,f; colormap)

Plot piecewise linear function on triangular grid given as "heatmap"

PlutoVista.triplot! — Method

 triplot!(p::VTKPlot,pts, tris,f)

Plot piecewise linear function on triangular grid given by points and triangles as matrices