Plotly API

This API is normally being used via PlutoVistaPlot.

PlutoVista.PlutoPlotlyPlot — Type

mutable struct PlutoPlotlyPlot <: PlutoVista.AbstractPlutoVistaBackend

Structure containing plot information for Plotly.js

PlutoVista.PlutoPlotlyPlot — Method

PlutoPlotlyPlot(; resolution, kwargs...) -> PlutoPlotlyPlot

Create a plotly plot.

PlutoVista.plot! — Method

plot!(p, x, y; kwargs...)

1D plotly.js plot

PlutoVista.tricontour! — Method

tricontour!(p, pts, tris, f; kwargs...)

Experimental. Plot piecewise linear function on triangular grid given as "heatmap" and with isolines using Plotly's mesh3d.

PlutoVista.contour! — Method

contour!(p, X, Y, f; kwargs...)

Experimental. Plot heatmap and isolines on rectangular grid defined by X and Y using Plotly's native contour plot.

PlutoVista.triplot! — Method

triplot!(p, pts, tris, f)

Experimental. Plot piecewise linear function on triangular grid given by points and triangles as matrices