vtk.js API

This API is normally being used via PlutoVistaPlot.

PlutoVista.PlutoVTKPlot — Type

mutable struct PlutoVTKPlot <: PlutoVista.AbstractPlutoVistaBackend

Structure containig plot information for vtk.js.

PlutoVista.PlutoVTKPlot — Method

PlutoVTKPlot(; resolution, kwargs...)

Create a vtk plot with given resolution in the notebook.

PlutoVista.tricontour! — Method

tricontour!(p, pts, tris, f; kwargs...)

Plot piecewise linear function on triangular grid given as "heatmap". Isolines can be given as a number or as a range.

PlutoVista.quiver2d! — Method

quiver2d!(p, pts, qvec; kwargs...)

2D quiver.

PlutoVista.triplot! — Method

triplot!(p, pts, tris, f; kwargs...)

Experimental: Plot piecewise linear function on triangular grid given by points and triangles as matrices

PlutoVista.contour! — Method

contour!(p, X, Y, f; kwargs...)

Plot piecewise linear function on triangular grid created from the tensor product of X and Y arrays as "heatmap". Levels can be given as a number or as a range.

PlutoVista.trimesh! — Method

trimesh!(p, pts, tris; kwargs...)

Plot triangular grid with optional region and boundary markers.

PlutoVista.tetcontour! — Method

tetcontour!(p, pts, tets, func; parentpts, kwargs...)

Plot isosurfaces given by levels and contour maps on planes given by the *planes parameters for piecewise linear function on tetrahedral mesh.

pts,tets,func can be arrays of grid data, allowing to plot discontinouos functions

PlutoVista.tetmesh! — Method

tetmesh!(p, pts, tets; kwargs...)

Plot parts of tetrahedral mesh below the planes given by the *plane parameters.